Get dialog window working
Dialog window only shows a label

Let's make it work as if it were the console.

This has created a ton of style errors, but I am too fatigued and too overdue to fix it. So it's done.
