--- src/main/java/vecrosen/Launcher.java
@@ -1,9 +1,0 @@
-package vecrosen;
-
-import javafx.application.Application;
-
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
