--- src/main/java/vecrosen/MainWindow.java
@@ -1,37 +1,0 @@
-package vecrosen;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.VBox;
-
-public class MainWindow extends AnchorPane {
-    @FXML
-    private ScrollPane scrollPane;
-    @FXML
-    private VBox dialogContainer;
-    @FXML
-    private TextField userInput;
-    @FXML
-    private Button sendButton;
-
-    private Vecrosen vecrosen;
-
-    @FXML
-    private void handleUserInput() {
-        System.err.println("Hello");
-        String input = userInput.getText();
-        //String response = vecrosen.getResponse(input);
-        System.err.println("World");
-        /*dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(input, userImage),
-                DialogBox.getDukeDialog(response, dukeImage)
-        );//*/
-        userInput.clear();
-    }
-
-
-}
